fix: don't close archive when FileObject streams close (#6)
Currently, if you open an input stream for a file and then close it,
the entire archive is closed, but VFS will keep trying to use it, and
any future input streams you open will just close immediately (because
junrar silently swallows an NPE in the pipe writer thread).

The built-in providers have this same notifyAllStreamsClosed() method
commented out. It does not seem like correct behavior.